sua bieu do, th, bat loi xoa trong dh

+++ webBanDienGiaDung/src/main/java/ptithcm/controller/adminController.java
@@ -14,7 +14,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
 
-import org.hibernate.cfg.annotations.MapKeyColumnDelegator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -68,9 +67,9 @@ public class adminController {
 		model.addAttribute("tongDoanhThu", tongDoanhThu);
 		
 		// Tính tổng doanh thu theo từng tháng
-	    List<Integer> monthlyRevenues = new ArrayList<>();
+	    List<Long> monthlyRevenues = new ArrayList<>();
 	    for (int i = 1; i <= 12; i++) {
-	        int totalRevenue = donHangService.tinhTongDoanhThuTheoThang(i);
+	        long totalRevenue = donHangService.tinhTongDoanhThuTheoThang(i);
 	        monthlyRevenues.add(totalRevenue);
 	    }
